refactor: move things around
- use fts_terms and index_terms in sql instead of adding new fields in SearchRequest proto
- move fst index search functions to src/service/search/grpc/storage.rs
- move file name converter function to config utils
