+++ src/job/files/idx.rs
@@ -16,8 +16,8 @@
 use config::{
     ider,
     meta::stream::{FileMeta, StreamType},
-    utils::parquet::new_parquet_writer,
-    FILE_EXT_IDX, FILE_EXT_PARQUET,
+    utils::{inverted_index::convert_parquet_idx_file_name, parquet::new_parquet_writer},
+    FILE_EXT_PARQUET,
 };
 use infra::storage;
 use tokio::task;
@@ -39,40 +39,6 @@ fn generate_index_file_name_from_compacted_file(
     );
     let file_name = ider::generate();
     format!("files/{stream_key}/{file_date}/{file_name}{FILE_EXT_PARQUET}")
-}
-
-/// FST inverted index solution has a 1:1 mapping between parquet and idx files.
-/// This is a helper function to convert the file names between the two by changing
-/// both the stream_type and extension parts of the file_name
-/// e.g.
-/// from: files/default/logs/quickstart1/2024/02/16/16/7164299619311026293.parquet
-/// to:   files/default/index/quickstart1/2024/02/16/16/7164299619311026293.idx
-/// and vice versa
-pub fn convert_parquet_idx_file_name(from: &str) -> String {
-    let mut parts: Vec<&str> = from.split('/').collect();
-    let stream_type_index = parts
-        .iter()
-        .position(|part| *part == "logs" || *part == "index")
-        .unwrap();
-
-    // Replace the stream_type part
-    parts[stream_type_index] = if parts[stream_type_index] == "logs" {
-        "index"
-    } else {
-        "logs"
-    };
-
-    // Replace the file extension
-    let file_name_index = parts.len() - 1;
-    let file_name = parts[file_name_index];
-    let new_file_name = if file_name.ends_with(FILE_EXT_PARQUET) {
-        file_name.replace(FILE_EXT_PARQUET, FILE_EXT_IDX)
-    } else {
-        file_name.replace(FILE_EXT_IDX, FILE_EXT_PARQUET)
-    };
-    parts[file_name_index] = &new_file_name;
-
-    parts.join("/")
 }
 
 pub(crate) async fn write_fst_index_to_disk(
@@ -190,28 +156,6 @@ pub(crate) async fn write_parquet_index_to_disk(
         Err(err) => {
             log::error!("[JOB] disk file upload error: {:?}", err);
             Err(anyhow::anyhow!(err))
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_convert_parquet_idx_file_name() {
-        let test_cases = vec![
-            (
-                "files/default/logs/quickstart1/2024/02/16/16/7164299619311026293.parquet",
-                "files/default/index/quickstart1/2024/02/16/16/7164299619311026293.idx",
-            ),
-            (
-                "files/default/index/quickstart1/2024/02/16/16/7164299619311026293.idx",
-                "files/default/logs/quickstart1/2024/02/16/16/7164299619311026293.parquet",
-            ),
-        ];
-        for (input, expected) in test_cases {
-            assert_eq!(convert_parquet_idx_file_name(input), expected);
         }
     }
 }
